refactor: separate resource code intel
